# Use bulk delete endpoint