This PR changes the endpoints to use a new, more optimised endpoint for bulk deletion. The API now makes sure to do a single SQL query to delete 1 or many entries, avoiding N+1 issue and improving overall perfomance, in both memory allocation and response speed

Before (just a snippet, this is actually 5x longer trace...)
```ruby
Started DELETE "/api/v1/manga_entries/7303" for ::1 at 2019-12-03 13:26:32 +0000
Processing by Api::V1::MangaEntriesController#destroy as HTML
  Parameters: {"id"=>"7303"}
  MangaEntry Load (0.4ms)  SELECT "manga_entries".* FROM "manga_entries" WHERE "manga_entries"."id" = $1 LIMIT $2  [["id", 7303], ["LIMIT", 1]]
  ↳ app/controllers/api/v1/manga_entries_controller.rb:67:in `destroy'
  MangaList Load (0.3ms)  SELECT "manga_lists".* FROM "manga_lists" WHERE "manga_lists"."id" = $1 LIMIT $2  [["id", 2], ["LIMIT", 1]]
  ↳ app/controllers/api/v1/manga_entries_controller.rb:71:in `destroy'
  User Load (0.3ms)  SELECT "users".* FROM "users" WHERE "users"."id" = $1 LIMIT $2  [["id", 1], ["LIMIT", 1]]
  ↳ app/controllers/api/v1/manga_entries_controller.rb:71:in `destroy'
  CACHE User Load (0.0ms)  SELECT "users".* FROM "users" WHERE "users"."id" = $1 LIMIT $2  [["id", 1], ["LIMIT", 1]]
  ↳ app/controllers/application_controller.rb:10:in `current_user'
   (0.3ms)  BEGIN
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
  MangaEntry Destroy (0.9ms)  DELETE FROM "manga_entries" WHERE "manga_entries"."id" = $1  [["id", 7303]]
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
   (5.0ms)  COMMIT
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
Completed 200 OK in 19ms (Views: 0.1ms | ActiveRecord: 7.2ms | Allocations: 6387)

Started DELETE "/api/v1/manga_entries/7293" for ::1 at 2019-12-03 13:26:32 +0000
   (2.7ms)  BEGIN
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
   (6.7ms)  COMMIT
  MangaEntry Destroy (4.9ms)  DELETE FROM "manga_entries" WHERE "manga_entries"."id" = $1  [["id", 7299]]
   (7.4ms)  COMMIT
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
Completed 200 OK in 73ms (Views: 0.3ms | ActiveRecord: 21.2ms | Allocations: 24178)

  MangaEntry Destroy (4.0ms)  DELETE FROM "manga_entries" WHERE "manga_entries"."id" = $1  [["id", 7296]]
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
Processing by Api::V1::MangaEntriesController#destroy as HTML
   (1.2ms)  COMMIT
  Parameters: {"id"=>"7293"}
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
  ↳ app/controllers/api/v1/manga_entries_controller.rb:75:in `destroy'
Started DELETE "/api/v1/manga_entries/7294" for ::1 at 2019-12-03 13:26:32 +0000
Completed 200 OK in 70ms (Views: 0.1ms | ActiveRecord: 31.0ms | Allocations: 24642)
```

After
```ruby
  MangaEntry Destroy (19.4ms)  DELETE FROM "manga_entries" WHERE "manga_entries"."id" IN ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15)  [["id", 7202], ["id", 7201], ["id", 7204], ["id", 7198], ["id", 7197], ["id", 7205], ["id", 7199], ["id", 7195], ["id", 7196], ["id", 7203], ["id", 7191], ["id", 7192], ["id", 7200], ["id", 7193], ["id", 7194]]
```

This also now shows how many entries have been deleted and properly handles if there was an error during update.

![new-delete-preview](https://user-images.githubusercontent.com/4270980/70055331-3af98280-15d1-11ea-8cd5-095c30a01f97.gif)


![delete-failed](https://user-images.githubusercontent.com/4270980/70055205-0259a900-15d1-11ea-90bd-598aa9e45d99.gif)